fix : DBcontroller DB query 삭제
